+++ src/middleware.ts
@@ -1,0 +1,5 @@
+import { SubscriptionOptions } from './client';
+
+export interface MiddlewareInterface {
+  applyMiddleware(options: SubscriptionOptions, next: Function): void;
+}