+++ src/test/tests.ts
@@ -418,6 +418,73 @@ describe('Client', function () {
     });
   });
 
+  it('should override SubscriptionOptions with middleware', function (done) {
+    const CTX = 'testContext';
+    const CTX2 = 'overrideContext';
+    const client3 = new SubscriptionClient(`ws://localhost:${TEST_PORT}/`);
+    client3.use([{
+      applyMiddleware(opts, next) {
+        // modify options for SubscriptionClient.subscribe here
+        opts.context = CTX2;
+        next();
+      },
+    }]);
+
+    client3.subscribe({
+        query: `subscription context {
+          context
+        }`,
+        variables: {},
+        context: CTX,
+      }, (error: any, result: any) => {
+        client3.unsubscribeAll();
+        if (error) {
+          assert(false);
+        }
+        if (result) {
+          assert.property(result, 'context');
+          assert.equal(result.context, CTX2);
+        }
+        done();
+      }
+    );
+    setTimeout(() => {
+      subscriptionManager.publish('context', {});
+    }, 100);
+  });
+
+  it('should override SubscriptionOptions with onSubscribe', function (done) {
+    const CTX = 'testContext';
+    const CTX2 = 'overrideContext';
+    const client3 = new SubscriptionClient(`ws://localhost:${TEST_PORT}/`);
+
+    client3.onSubscribe((opts) => {
+      opts.context = CTX2;
+    });
+
+    client3.subscribe({
+        query: `subscription context {
+          context
+        }`,
+        variables: {},
+        context: CTX,
+      }, (error: any, result: any) => {
+        client3.unsubscribeAll();
+        if (error) {
+          assert(false);
+        }
+        if (result) {
+          assert.property(result, 'context');
+          assert.equal(result.context, CTX2);
+        }
+        done();
+      }
+    );
+    setTimeout(() => {
+      subscriptionManager.publish('context', {});
+    }, 100);
+  });
+
   it('should handle correctly init_fail message', (done) => {
     wsServer.on('connection', (connection: any) => {
       connection.on('message', (message: any) => {